Add files via upload
My best current version. My best results came when going travelling the total route with the lowest Manhattan distance, similar to your idea. 

I've tried to implement it in a way which sets the pick-up as the (cost) nearest one instead of the (distance) nearest one, but this did not improve my results so eventually I also tried computing the total Manhattan distance and this did improve results significantly.