My best current version
My best results also come when I travel the shortest total route (Manhattan). I also successfully implemented a version in which the next package to be picked up is always the (cost) nearest one to the current position but this did not improve my results compared to the one which went for the (distance) nearest one.